Add major edits to the Html/css
I made all of the html structure, trying to minimize non-semantic divs, etc. The CSS was an attempt to make a two colomn layout. One thing I know is messed up is the width of the project box, because it doesn't seem to scale at the same rate as the other two boxes above it.